+++ internal/store/types.go
@@ -23,12 +23,14 @@ func (c Chat) Muted() bool {
 }
 
 type Group struct {
-	JID       string
-	Name      string
-	OwnerJID  string
-	CreatedAt time.Time
-	LeftAt    time.Time
-	UpdatedAt time.Time
+	JID             string
+	Name            string
+	OwnerJID        string
+	CreatedAt       time.Time
+	IsParent        bool
+	LinkedParentJID string
+	LeftAt          time.Time
+	UpdatedAt       time.Time
 }
 
 type GroupParticipant struct {